requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai>=1.14.0
python-dotenv==1.0.0
streamlit==1.30.0
requests==2.31.0
pandas>=1.5.0 